+ updated the WebSocketServer transport to represent a single connection to the server + updated Node.js example project to utilize modified transport
