+++ lib/node/src/transport/WebSocketServer.js
@@ -31,15 +31,13 @@
 */
 
 const { TransportBase } = require('./TransportBase.js');
-const { SslConfig } = require('./SslConfig.js');
 const { SdlPsm } = require('./SdlPsm.js');
 const WebSocket = require('ws');
-const https = require('https');
 
 class WebSocketServer extends TransportBase {
     constructor (transportConfig, transportCallback = null) {
         super(transportConfig, transportCallback);
-        this._ws = null;
+        this._connectionLostInterval = null;
     }
 
     /**
@@ -47,92 +45,65 @@ class WebSocketServer extends TransportBase {
      */
     setTransportCallback (callback) {
         this._transportCallback = callback;
+        return this;
     }
 
     /**
-     * Start listening for connections to the WebSocket Server
+     * Start listening for events from the client
      * @return {WebSocketServer}
      */
     start () {
-        if (
-            this._transportConfig.getSslConfig() instanceof SslConfig
-            && this._transportConfig.getSslConfig().getPemCertificate() !== null
-            && this._transportConfig.getSslConfig().getPrivateKey() !== null
-        ) {
-            // create a WebSocket Secure Server
-            const server = https.createServer({
-                cert: this._transportConfig.getSslConfig().getPemCertificate(),
-                key: this._transportConfig.getSslConfig().getPrivateKey(),
-                passphrase: this._transportConfig.getSslConfig().getPassword(),
-            });
-            this._ws = new WebSocket.Server({
-                server,
-            });
-            server.listen(this._transportConfig.getPort());
-            console.log(`WSS started on port ${this._transportConfig.getPort()}`);
-        } else {
-            // create a WebSocket Server
-            this._ws = new WebSocket.Server({
-                port: this._transportConfig.getPort(),
-            });
-            console.log(`WS started on port ${this._transportConfig.getPort()}`);
-        }
+        const client = this._transportConfig.getClient();
+        const connectionLostTimeout = this._transportConfig.getConnectionLostTimeout();
 
-        // Event listener for incoming WebSocket connections
-        this._ws.on('connection', (connection) => {
-            console.log('connection established');
-
-            // Event listener for an incoming message
-            connection.on('message', (message) => {
-                this._handleMessage(message);
-            });
-
-            // Event listener for a closed connection
-            connection.on('close', () => {
-                console.log('server received close event');
-                if (this._transportCallback !== null) {
-                    this._transportCallback.onConnectionTerminated();
-                }
-            });
-
-            // Event listener for errors
-            connection.on('error', (errorEvent) => {
-                console.log('server received error event');
-                if (this._transportCallback !== null) {
-                    this._transportCallback.onError();
-                }
-            });
-
-            // Pong heartbeat listener
-            connection.on('pong', () => {
-                console.log('server received pong event');
-                connection.isAlive = true;
-            });
-
-            // Trigger event for connection established
-            // FIXME: pass a client context to the transport listener?
-            if (this._transportCallback !== null) {
-                this._transportCallback.onConnectionEstablished();
-            }
-            connection.isAlive = true;
+        // Event listener for an incoming message
+        client.on('message', (message) => {
+            this._handleMessage(message);
         });
 
-        // Detect broken connections
-        if (this._transportConfig.getConnectionLostTimeout() > 0) {
-            setInterval(() => {
-                this._ws.clients.forEach(function each (client) {
-                    if (client.isAlive === false) {
-                        console.log('server deemed client dead');
-                        return client.terminate();
-                    }
+        // Event listener for a closed connection
+        client.on('close', () => {
+            console.log('server received close event');
+            if (this._transportCallback !== null) {
+                this._transportCallback.onConnectionTerminated();
+            }
+        });
 
-                    // Assume the client is dead and ask it if it's alive
-                    client.isAlive = false;
-                    console.log('server sending ping to client');
-                    client.ping(function () {});
-                });
-            }, this._transportConfig.getConnectionLostTimeout());
+        // Event listener for errors
+        client.on('error', (errorEvent) => {
+            console.log('server received error event');
+            if (this._transportCallback !== null) {
+                this._transportCallback.onError();
+            }
+        });
+
+        // Pong heartbeat listener
+        client.on('pong', () => {
+            console.log('server received pong event');
+            client.isAlive = true;
+        });
+
+        // Trigger event for connection established
+        if (this._transportCallback !== null) {
+            this._transportCallback.onConnectionEstablished();
         }
+
+        // Detect broken connections
+        if (connectionLostTimeout > 0) {
+            this._connectionLostInterval = setInterval(() => {
+                if (client.isAlive === false) {
+                    console.log('server deemed client dead');
+                    return client.terminate();
+                }
+
+                // Assume the client is dead and ask it if it's alive
+                client.isAlive = false;
+                console.log('server sending ping to client');
+                client.ping(function () {});
+            }, connectionLostTimeout);
+        }
+
+        client.isAlive = true;
 
         return this;
     }
@@ -142,25 +113,25 @@ class WebSocketServer extends TransportBase {
      * @return {WebSocketServer}
      */
     stop () {
-        if (this._ws !== null) {
-            this._ws.close();
+        const client = this._transportConfig.getClient();
+
+        if (this._connectionLostInterval !== null) {
+            clearInterval(this._connectionLostInterval);
         }
+
+        client.terminate();
 
         return this;
     }
 
     /**
-     * Sends a packet to all connected WebSocket clients
-     * FIXME: Sending packets should be isolated to a client rather than targeting all clients
+     * Sends a packet to the connected WebSocket Server client
      * @return {WebSocketServer}
      */
     sendPacket (sdlPacket) {
-        if (this._ws !== null) {
-            this._ws.clients.forEach(function each (client) {
-                if (client.readyState === WebSocket.OPEN) {
-                    client.send(sdlPacket.toPacket());
-                }
-            });
+        const client = this._transportConfig.getClient();
+        if (client.readyState === WebSocket.OPEN) {
+            client.send(sdlPacket.toPacket());
         }
 
         return this;
