+++ lib/node/src/transport/WebSocketServerConfig.js
@@ -36,22 +36,24 @@ const { TransportType } = require('./enums/TransportType.js');
 class WebSocketServerConfig extends TransportConfigBase {
     /**
     * @constructor
-    * @param {Number} port - The port to listen for WebSocket connections on.
-    * @param {Number} connectionLostTimeout - The timeout for a connection lost, represented in milliseconds. Default 60000. If a value less than 0 is used, then the websocket will wait indefinitely.
+    * @param {WebSocket.client} client - A client connection reference made from a WebSocket server
     */
-    constructor (port = 3000, connectionLostTimeout = 60000, sslConfig = null) {
+    constructor (client = null, connectionLostTimeout = 60000) {
         super(TransportType.WEBSOCKET_SERVER);
-        this._port = port;
+        this._client = client;
         this._connectionLostTimeout = connectionLostTimeout;
-        this._sslConfig = sslConfig;
+
+        if (client === null) {
+            throw Error('client must be a WebSocket client');
+        }
     }
 
     /**
-    * Returns the websocket listener port
-    * @return {Number}
+    * Returns the websocket client
+    * @return {WebSocket.client}
     */
-    getPort () {
-        return this._port;
+    getClient () {
+        return this._client;
     }
 
     /**
@@ -60,14 +62,6 @@ class WebSocketServerConfig extends TransportConfigBase {
     */
     getConnectionLostTimeout () {
         return this._connectionLostTimeout;
-    }
-
-    /**
-    * Returns the SSL configuration
-    * @return {SslConfig}
-    */
-    getSslConfig () {
-        return this._sslConfig;
     }
 }
